backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.12
twilio==8.5.0
python-multipart
